refactor blueprints toolbar into separate component
