+++ src/options/pages/blueprints/BlueprintsCard.tsx
@@ -51,6 +51,7 @@ import { useSelector } from "react-redux";
 import { uniq } from "lodash";
 import useInstallableViewItems from "@/options/pages/blueprints/useInstallableViewItems";
 import AutoSizer from "react-virtualized-auto-sizer";
+import BlueprintsToolbar from "@/options/pages/blueprints/BlueprintsToolbar";
 
 // These react-table columns aren't rendered as column headings,
 // but used to expose grouping, sorting, filtering, and global
@@ -110,17 +111,14 @@ const BlueprintsCard: React.FunctionComponent<{
     [data]
   );
 
-  const [view, setView] = useReduxState(
-    selectView,
-    blueprintsSlice.actions.setView
-  );
+  const [view] = useReduxState(selectView, blueprintsSlice.actions.setView);
 
-  const [groupBy, setGroupBy] = useReduxState(
+  const [groupBy] = useReduxState(
     selectGroupBy,
     blueprintsSlice.actions.setGroupBy
   );
 
-  const [sortBy, setSortBy] = useReduxState(
+  const [sortBy] = useReduxState(
     selectSortBy,
     blueprintsSlice.actions.setSortBy
   );
@@ -154,37 +152,7 @@ const BlueprintsCard: React.FunctionComponent<{
     useSortBy
   );
 
-  const {
-    rows,
-    flatRows,
-    flatHeaders,
-    setGlobalFilter,
-    state: { globalFilter },
-  } = tableInstance;
-
-  const isGrouped = groupBy.length > 0;
-  const isSorted = sortBy.length > 0;
-  const numberOfBlueprints = isGrouped
-    ? flatRows.length - rows.length
-    : rows.length;
-
-  const { groupByOptions, sortByOptions } = useMemo(() => {
-    const groupByOptions = flatHeaders
-      .filter((column) => column.canGroupBy)
-      .map((column) => ({
-        label: column.Header,
-        value: column.id,
-      }));
-
-    const sortByOptions = flatHeaders
-      .filter((column) => column.canSort)
-      .map((column) => ({
-        label: column.Header,
-        value: column.id,
-      }));
-
-    return { groupByOptions, sortByOptions };
-  }, [flatHeaders]);
+  const { rows, setGlobalFilter } = tableInstance;
 
   const BlueprintsView = view === "list" ? ListView : GridView;
 
@@ -195,82 +163,7 @@ const BlueprintsCard: React.FunctionComponent<{
         setGlobalFilter={setGlobalFilter}
       />
       <Col className={styles.mainContainer}>
-        <div className="d-flex justify-content-between align-items-center mb-3">
-          <h3 className={styles.filterTitle}>
-            {globalFilter
-              ? `${numberOfBlueprints} results for "${globalFilter}"`
-              : `${filters.length > 0 ? filters[0].value : "All"} Blueprints`}
-          </h3>
-          <span className="d-flex align-items-center small">
-            <span className="ml-3 mr-2">Group by:</span>
-            <Select
-              isClearable
-              placeholder="Group by"
-              options={groupByOptions}
-              onChange={(option, { action }) => {
-                const value = action === "clear" ? [] : [option.value];
-                setGroupBy(value);
-              }}
-              value={groupByOptions.find((opt) => opt.value === groupBy[0])}
-            />
-
-            <span className="ml-3 mr-2">Sort by:</span>
-            <Select
-              isClearable
-              placeholder="Sort by"
-              options={sortByOptions}
-              onChange={(option, { action }) => {
-                const value =
-                  action === "clear" ? [] : [{ id: option.value, desc: false }];
-                setSortBy(value);
-              }}
-              value={sortByOptions.find((opt) => opt.value === sortBy[0]?.id)}
-            />
-
-            {isSorted && (
-              <Button
-                variant="link"
-                size="sm"
-                onClick={() => {
-                  const value = [{ id: sortBy[0].id, desc: !sortBy[0].desc }];
-                  setSortBy(value);
-                }}
-              >
-                <FontAwesomeIcon
-                  icon={
-                    sortBy[0].id === "updatedAt"
-                      ? sortBy[0].desc
-                        ? faSortAmountDownAlt
-                        : faSortAmountUpAlt
-                      : sortBy[0].desc
-                      ? faSortAmountUpAlt
-                      : faSortAmountDownAlt
-                  }
-                  size="lg"
-                />
-              </Button>
-            )}
-            <Button
-              variant={view === "list" ? "link" : "outline-link"}
-              size="sm"
-              className="ml-3"
-              onClick={() => {
-                setView("list");
-              }}
-            >
-              <FontAwesomeIcon icon={faList} size="lg" />
-            </Button>
-            <Button
-              variant={view === "grid" ? "link" : "outline-link"}
-              size="sm"
-              onClick={() => {
-                setView("grid");
-              }}
-            >
-              <FontAwesomeIcon icon={faThLarge} size="lg" />
-            </Button>
-          </span>
-        </div>
+        <BlueprintsToolbar tableInstance={tableInstance} />
         {/* This wrapper prevents AutoSizer overflow in a flex box container */}
         <div style={{ flex: "1 1 auto" }}>
           <AutoSizer defaultHeight={500}>
